requirements.txt
numpy
streamlit
openai
langchain
PyMuPDF
python-dotenv
tiktoken
langchain-community
sentence-transformers